mm-radio: Move all frontend-related code into src/radio directory
Makes it easier to navigate through the code.
