+++ src/meson.build
@@ -1,0 +1,7 @@
+srcs += files(
+    'service.cpp',
+    'RadioHalManager.cpp',
+)
+
+subdir('mm')
+subdir('radio')
