+++ src/mm/meson.build
@@ -1,0 +1,6 @@
+srcs += files(
+    'Modem.cpp',
+    'ModemManager.cpp',
+    'ModemVoice.cpp',
+    'ModemUssd.cpp',
+)
