+++ src/radio/meson.build
@@ -1,0 +1,9 @@
+srcs += files(
+    'RadioConfig.cpp',
+    'RadioData.cpp',
+    'RadioMessaging.cpp',
+    'RadioModem.cpp',
+    'RadioNetwork.cpp',
+    'RadioSim.cpp',
+    'RadioVoice.cpp',
+)
